Completed BoolPrompt implementation
BoolPrompt is now fully implemented, and controls and draws appropriately (though design itself might change).

There is also now also a universal QuitPrompt, which is effectively a wrapper for a bool prompt with standardized arguments.

This commit also includes the introduction of a new `terminalgui` module, that combines write and draw specifically for interface elements.