Reworked the game module and added a dungeon generator
Game is now reserved for setting up and cleaning up the game. Dungeon now contains generation of the scene. The main menu is now in charge of calling for a new game. And Main is now responsible for initializing modules.

This adds a little system complexity, but responsibility is clearer.

Launching the app now starts the game in the main menu, where you can start a new game, and play as before.