Added sprite layers and sorting
It is now possible to control which sprite gets rendered on top of another sprite!

Performs well enough for the small amount of sprites used (currently 16), but should be reviewed for large amounts. The priority sorting, for example, should probably happen when setting the priority, instead of on every call to render.