Added inventory view and fix input alt keys
Input alt keys were all set to the wrong input action. This has now been fixed. Toggle for pause menu has also been removed as an input, as the expectation is that it appears when there is no more menus to exit out of, while the game is in play.

Added a very primitive inventory view to check collected items, once implemented.