Reversed world rendering enabled logic and added option for disabling turn processes
Previously, `require` allowed for the ability to turn on rendering (and now turn processing), but did not allow for the option of disabling it.

This caused issues for views like the inventory, which, because it was pushed onto the view stack, had the game default view override its desire to disable the world rendering.

Strangely, this reverse in logic allows behavior as originally intended.